+++ src/language-server/__tests__/fixtures/starwarsSchema.ts
@@ -330,6 +330,7 @@ export const starwarsSchema: IntrospectionQuery = {
             name: "Droid",
           },
         ],
+        interfaces: [],
       },
       {
         kind: "SCALAR",