+++ src/language-server/typings/codemirror.d.ts
@@ -1,0 +1,4 @@
+// `codemirror` types have some issues, and this isn't really being used anyways
+declare module "codemirror" {
+  type Token = any;
+}