+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Specify files with conflicting documents for the 'There are multiple definitions' message. [#29](https://github.com/apollographql/vscode-graphql/pull/29)
 - Dont watch non file documents. Diff view will no longer crash extension. [#28](https://github.com/apollographql/vscode-graphql/pull/28)
+- Upgrade graphql version to 15.5 [#34](https://github.com/apollographql/vscode-graphql/pull/34)
 
 ### 1.19.6